Refactored package
