+++ README.md
@@ -15,3 +15,29 @@ Example use cases:
 
 - https://github.com/lamastex/spark-texata-2020
 - 
+
+## Included parsers
+
+Parsers for 1-minute foreign exchange data (https://github.com/philipperemy/FX-1-Minute-Data) and stock market data from yfinance (https://github.com/ranaroussi/yfinance). The data from yfinance requires some processing in python before being accepted by the parser.
+
+### Foreign Exchange parser
+
+The scala function is `parseFX` and has as input a string formatted as 
+
+```
+"DateTime Stamp;Bar OPEN Bid Quote;Bar HIGH Bid Quote;Bar LOW Bid Quote;Bar CLOSE Bid Quote;Volume"
+```
+
+where DateTime Stamp is formatted as `yyyyMMdd HHmmSS`. Open, High, Low and Close are floating point numbers and Volume is an integer (which seems to always be 0).
+
+
+
+### Yahoo! Finance parser
+
+The scala function is `parseYF` and has as input a string formatted as
+
+```
+"DateTime Stamp,Open,High,Low,Close,Adj Close,Volume"
+```
+
+where DateTime Stamp is formatted either as `yyyy-MM-dd` or beginning with `yyyy-MM-dd HH:mm:SS` (i.e. `2020-07-09 18:05:00+02:00` is valid but `UTC+2 2020-07-09 18:05:00` is not). Open, High, Low, Close and Adj Close are floating point numbers and Volume can be either an integer or a floating point number.
