+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.lamastex.trendcalculus</groupId>
+    <groupId>org.lamastex.spark</groupId>
     <artifactId>TrendCalculus-2020</artifactId>
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
