+++ src/main/scala/org/lamastex/trendcalculus/Point.scala
@@ -1,4 +1,4 @@
-package org.lamastex.trendcalculus
+package org.lamastex.spark.trendcalculus
 
 case class Point(
                   x: Long,
