+++ src/main/scala/org/lamastex/trendcalculus/TrendCalculus.scala
@@ -12,9 +12,9 @@
  * limitations under the License.
  */
 
-package org.lamastex.trendcalculus
+package org.lamastex.spark.trendcalculus
 
-import org.lamastex.trendcalculus.DateUtils.Frequency
+import org.lamastex.spark.trendcalculus.DateUtils.Frequency
 
 import scala.util.Try
